# contracts/operator.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Max operator API",
    "version": "1.0.0",
...
  },
  "servers": [{ "url": "https://{host}/operator/v1", "variables": { "host": { "default": "api.example.com", "description": "This deployment's backend host" } } }],
  "security": [{ "serviceIdentity": [] }],
...
  "paths": {
    "/whoami": {
      "get": {
        "tags": ["console"],
        "operationId": "whoami",
        "summary": "Caller identity and build (smoke test)",
        "parameters": [{ "$ref": "#/components/parameters/actor" }],
        "responses": {
          "200": {
            "description": "Identity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "caller": {
                      "type": ["string", "null"],
                      "description": "The audit actor of this request: a valid X-Operator-Actor, else the caller's service-account e-mail ('dev' with the dev token)"
                    },
                    "build": { "type": ["string", "null"], "description": "BUILD_GIT_SHA of this deployment" }
                  },
                  "required": ["caller", "build"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/models": {
      "get": {
        "tags": ["console"],
        "operationId": "listModels",
        "summary": "This build's model registry",
        "description": "What this deployment can run, per model tier. A console or benchmark service cannot know it otherwise. Local-LLM ids are left out. Cache-Control: no-store.",
        "parameters": [{ "$ref": "#/components/parameters/actor" }],
        "responses": {
          "200": { "description": "Registry", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ModelCatalogue" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/audit": {
      "get": {
        "tags": ["audit"],
        "operationId": "listAudit",
        "summary": "Admin action audit trail, newest first",
        "parameters": [
          { "$ref": "#/components/parameters/actor" },
          { "$ref": "#/components/parameters/from" },
          { "$ref": "#/components/parameters/to" },
          { "name": "q", "in": "query", "description": "Case-insensitive substring of action, target_id, target_type or actor (trimmed)", "schema": { "type": "string" } },
          { "$ref": "#/components/parameters/limit" },
          { "$ref": "#/components/parameters/offset" }
        ],
        "responses": {
          "200": {
            "description": "Page",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "range": { "$ref": "#/components/schemas/Range" },
                    "limit": { "type": "integer" },
                    "offset": { "type": "integer" },
                    "total": { "type": "integer" },
                    "rows": { "type": "array", "items": { "$ref": "#/components/schemas/AuditRow" } }
                  },
                  "required": ["range", "limit", "offset", "total", "rows"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["audit"],
        "operationId": "recordAudit",
        "summary": "Record an action the console performed elsewhere",
        "description": "Keeps one admin_audit trail: the console logs actions it performed outside this API. actor = X-Operator-Actor or the caller. Once the body validates the answer is 201 — the insert is fire-and-forget like every audit write, so a database failure is logged, not returned.",
        "parameters": [{ "$ref": "#/components/parameters/actor" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": { "type": "string", "pattern": "^[a-z][a-z0-9_.-]{2,63}$", "description": "Dot-namespaced verb, e.g. console.user.note" },
                  "target_type": { "type": ["string", "null"], "maxLength": 200 },
                  "target_id": { "type": ["string", "null"], "maxLength": 200 },
                  "payload": { "type": ["object", "null"], "additionalProperties": true, "description": "At most 8 KiB as JSON; never secrets or document content" }
                },
                "required": ["action"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recorded",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "const": true } }, "required": ["ok"] } } }
          },
          "400": {
            "description": "action / target_type / target_id / payload invalid",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "413": { "description": "payload too large (max 8 KB)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/login-events": {
      "post": {
        "tags": ["console"],
        "operationId": "recordLogin",
        "summary": "An operator signed in to the console",
        "description": "Shifts the deployment-wide login pair (current → previous, current = now). 'previous' is the reference point of 'new users since last login' (GET /new-users, GET /users totals.new_users_count). Global, not per operator; shared with the legacy password login. No body; not audited.",
        "parameters": [{ "$ref": "#/components/parameters/actor" }],
        "responses": {
          "200": {
            "description": "Recorded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "const": true },
                    "previous_login_at": { "type": ["string", "null"], "format": "date-time", "description": "The login before this one; null on a fresh install" }
                  },
                  "required": ["ok", "previous_login_at"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/expiry-reminders/send": {
      "post": {
        "tags": ["jobs"],
        "operationId": "sendExpiryReminders",
        "summary": "Run the plan-expiry reminder sweep now (same as the cron)",
        "description": "Audited as email.expiry_reminders.send.",
        "parameters": [{ "$ref": "#/components/parameters/actor" }],
        "responses": {
          "200": {
            "description": "Sweep result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "const": true },
                    "expiring": { "type": "integer" },
                    "skipped_auto_renew": { "type": "integer" },
                    "already_reminded": { "type": "integer" },
                    "sent": { "type": "integer" }
                  },
                  "required": ["ok"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "serviceIdentity": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Google OIDC ID token",
        "description": "A Google-signed OIDC identity token of a service account (no shared password). Accepted when the signature verifies against Google's keys, email_verified is true and the email (case-insensitive) is listed in OPERATOR_API_CALLERS (separated by ',', ';' or whitespace — ';' survives gcloud's comma-split env flags), and aud is one of OPERATOR_API_AUDIENCE (same separators, trailing '/' ignored) or, when that is unset, this host's origin https://{Host} or https://{Host}/operator/v1. OPERATOR_API_CALLERS unset or empty ⇒ the API does not exist: every path answers 404 before the token is looked at. Outside production (NODE_ENV ≠ production) the static OPERATOR_API_DEV_TOKEN is accepted as well (audit actor fallback 'dev'), still only when OPERATOR_API_CALLERS is set."
      }
    },
    "parameters": {
      "actor": {
        "name": "X-Operator-Actor",
        "in": "header",
        "required": false,
        "description": "The human or tool behind this request (e.g. an operator's e-mail). Trimmed; anything not matching the pattern is ignored silently. Recorded as admin_audit.actor for every audited action and echoed by GET /whoami; absent or invalid → the caller's service-account e-mail ('dev' with the dev token). Trusted only because the caller itself is authenticated.",
        "schema": { "type": "string", "pattern": "^[\\w.@:+-]{1,100}$" }
      },
...
      "from": { "name": "from", "in": "query", "description": "Inclusive lower bound (ISO 8601); default now − 30 days", "schema": { "type": "string", "format": "date-time" } },
      "to": { "name": "to", "in": "query", "description": "Exclusive upper bound (ISO 8601); default now", "schema": { "type": "string", "format": "date-time" } },
      "limit": { "name": "limit", "in": "query", "description": "Page size; clamped to 1–500", "schema": { "type": "integer", "default": 50, "minimum": 1, "maximum": 500 } },
      "offset": { "name": "offset", "in": "query", "description": "Rows to skip; negative → 0", "schema": { "type": "integer", "default": 0, "minimum": 0 } },
...
    },
    "responses": {
      "Unauthorized": { "description": "Missing bearer token | Invalid caller token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "ServerError": {
        "description": "Database or upstream failure; detail carries the message",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["detail"],
        "properties": { "detail": { "type": "string" }, "error": { "type": "string", "description": "Exception message (global error handler only)" } },
        "additionalProperties": true
      },
...
      "Range": {
        "type": "object",
        "required": ["from", "to"],
        "properties": { "from": { "type": "string", "format": "date-time" }, "to": { "type": "string", "format": "date-time" } }
      },
...
      "ModelCatalogue": {
        "type": "object",
        "properties": {
          "defaults": { "type": "object", "properties": { "main": { "type": "string" }, "tabular": { "type": "string" }, "title": { "type": "string" } } },
          "tiers": {
            "type": "object",
            "properties": {
              "main": { "$ref": "#/components/schemas/ModelList" },
              "mid": { "$ref": "#/components/schemas/ModelList" },
              "low": { "$ref": "#/components/schemas/ModelList" }
            }
          },
          "build": { "type": ["string", "null"] }
        },
        "required": ["defaults", "tiers", "build"]
      },
      "ModelList": { "type": "array", "items": { "type": "object", "properties": { "id": { "type": "string" }, "provider": { "type": "string" } }, "required": ["id", "provider"] } },
...
      "AuditRow": {
        "type": "object",
        "properties": {
          "id": { "type": ["number", "string", "null"] },
          "actor": { "type": "string" },
          "action": { "type": "string" },
          "target_type": { "type": ["string", "null"] },
          "target_id": { "type": ["string", "null"] },
          "payload": { "type": ["object", "null"], "additionalProperties": true },
          "ip": { "type": ["string", "null"] },
          "created_at": { "type": "string", "format": "date-time" }
        },
        "additionalProperties": true
      },
...
    }
  }
}